requirements.txt
Django==3.1.5
django-svelte==0.1.6
djangorestframework==3.12.2
drf-yasg==1.20.0
